+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/AtomicAttribute2023.kt
@@ -12,20 +12,20 @@ package at.asitplus.wallet.lib.data
  * see the "LICENSE" file for more details
  */
 
+import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.decodeFromJsonElement
 
 /**
  * VC spec leaves the representation of a single credential open to implementations.
  * We decided to make a "generic" one, i.e. with custom [name], [value] and [mimeType].
  */
-@Suppress("DEPRECATION")
 @Serializable
-@SerialName("AtomicAttribute2023")
 data class AtomicAttribute2023(
-    override val id: String,
+    @SerialName("id")
+    val id: String,
 
     @SerialName("name")
     val name: String,
@@ -35,9 +35,21 @@ data class AtomicAttribute2023(
 
     @SerialName("mime-type")
     val mimeType: String,
-) : CredentialSubject() {
+) {
 
     constructor(id: String, name: String, value: String) : this(id, name, value, "application/text")
+
+    companion object {
+
+        /**
+         * Converts a [JsonElement] into a [AtomicAttribute2023] using kotlinx.serialization.
+         *
+         * @throws kotlinx.serialization.SerializationException if serialization fails
+         */
+        fun fromJsonElement(input: JsonElement): AtomicAttribute2023 =
+            joseCompliantSerializer.decodeFromJsonElement(input)
+
+    }
 
 }
 
@@ -48,4 +60,4 @@ data class AtomicAttribute2023(
  * @throws kotlinx.serialization.SerializationException if serialization fails
  */
 fun AtomicAttribute2023.toJsonElement(): JsonElement =
-    Json.encodeToJsonElement(AtomicAttribute2023.serializer(), this)
+    joseCompliantSerializer.encodeToJsonElement(AtomicAttribute2023.serializer(), this)