+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 Release 6.0.0 (unreleased):
- - Remove code deprecated in 5.12.0
  - Deprecations:
+    - Remove code deprecated in 5.12.0, e.g. `CredentialSubject` as base class for JWT VC
    - Deprecate `vckJsonSerializer`, should be replaced with `joseCompliantSerializer` (Signum)
 
 Release 5.12.0: